backend/requirements.txt
fastapi
uvicorn
ultralytics
firebase-admin
python-multipart
opencv-python-headless
numpy
pydantic
python-dotenv
aiofiles
